bus-screen-text-detector/micro-service/src/requirements.txt
uvicorn==0.19.0
fastapi==0.85.0
python-multipart==0.0.5
numpy==1.24.2
prometheus-fastapi-instrumentator==6.0.0
codecarbon==2.2.2
tqdm==4.65.0
minio==7.1.14
donut-python==1.0.9
transformers==4.25.1
timm==0.5.4
requests==2.29.0
pytorch-lightning==1.6.4
Pillow==9.5.0
